Move deprecated to another file.
